refactor(footer): refactor code by refer pr review
