+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "ng-homeland",
-  "version": "1.0.10",
+  "version": "1.0.11",
   "private": true,
   "scripts": {
     "dev": "next dev",
